Added ResponseEntity

+++ src/main/java/com/mentortheyoung/moneyflow/controllers/UserController.java
@@ -2,6 +2,8 @@ package com.mentortheyoung.moneyflow.controllers;
 
 import com.mentortheyoung.moneyflow.entities.User;
 import com.mentortheyoung.moneyflow.services.UserService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,31 +21,46 @@ public class UserController {
     }
 
     @GetMapping("/{userId}")
-    public User getUserById(@PathVariable Integer userId) {
-    logger.info("Getting user by id: " + userId);
-    return userService.getUserById(userId);
+    public ResponseEntity<User> getUserById(@PathVariable Integer userId) {
+        logger.info("Getting user by id: " + userId);
+        User user = userService.getUserById(userId);
+        if (user != null)
+            return new ResponseEntity<>(user, HttpStatus.OK);
+        else
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping
-    public List<User> getUsers() {
-        return userService.getUsers();
+    public ResponseEntity<List<User>> getUsers() {
+        return new ResponseEntity<>(userService.getUsers(), HttpStatus.OK);
     }
 
     @PostMapping
-    public User createUser(@RequestBody User user) {
+    public ResponseEntity <User> createUser(@RequestBody User user) {
         logger.info("Creating new user");
-        return userService.saveUser(user);
+        User savedUser = userService.saveUser(user);
+        return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
     }
 
     @PutMapping("/{userId}")
-    public User updateUser(@PathVariable Integer userId, @RequestBody User user) {
+    public ResponseEntity<String> updateUser(@PathVariable Integer userId, @RequestBody User user) {
         logger.info("Updating user with id: " + userId);
-        return userService.updateUser(userId, user);
+        User user1 = userService.updateUser(userId, user);
+        if (user1 != null)
+            return new ResponseEntity<>("Updated", HttpStatus.OK);
+        else
+            return new ResponseEntity<>("Failed to update", HttpStatus.BAD_REQUEST);
     }
 
     @DeleteMapping("/{userId}")
-    public  void  deleteUser(@PathVariable Integer userId) {
+    public ResponseEntity<String> deleteUser(@PathVariable Integer userId) {
         logger.info("Deleting user with id: " + userId);
-        userService.deleteUserById(userId);
+        User user = userService.getUserById(userId);
+        if (user != null) {
+            userService.deleteUserById(userId);
+            return new ResponseEntity<>("Deleted", HttpStatus.OK);
+        }
+        else
+            return new ResponseEntity<>("User not found", HttpStatus.NOT_FOUND);
     }
 }
